backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
cryptography==41.0.7
PyJWT==2.8.0
requests==2.31.0
Flask-Limiter==3.5.0
reportlab==4.0.7
Pillow==10.1.0
